Fixed #28401 - Allow hashlib.md5() calls to work with FIPS kernels
md5 is not an approved algorithm in FIPS mode, and trying to
instantiate a hashlib.md5() will fail when the system is running in
FIPS mode.

md5 is allowed when in a non-security context.  There is a plan to
add a keyword parameter (usedforsecurity) to hashlib.md5() to annotate
whether or not the instance is being used in a security context.

In the case where it is not, the instantiation of md5 will be allowed.
See https://bugs.python.org/issue9216 for more details.

Some downstream python versions already support this parameter.  To
support these versions, a new encapsulation of md5() has been added to
django/utils/crypto.py.  This encapsulation will pass through the
usedforsecurity parameter in the case where the parameter is supported,
and strip it if it is not.

In django, it appears that md5() is mostly used to generate cache keys
and file/db record names.  These are not being used in a security
context and should be allowed.  The md5() usages for the password
hashers, on the other hand, should not.

This is built on 2.2.4 tag - which is being used in openstack only as
a test.  The expectation is that openstack will move (maybe) to a
later version of django.